requirements.txt
openai
kivy
python-dotenv
Django==2.2.5